clutter/gesture: Introduce inhibition of gestures
Add a second way for gestures to influence each other: Failure requirements
aka inhibiting the start of other gestures.

Since this now allows for very complex interactions and possibly even
circular dependencies between gestures, this commit also makes a change to
the state-change debugging messages: For every recursive state-change, a few
spaces are added before the debug message so that the recursion can be
picked up easier by the eye.