clutter/stage: Cancel implicit grab if the whole grab tree goes unmapped
When an actor in the implicit grab tree goes unmapped, we normally just move
up the tree to the next mapped parent and use that as new implicit grab
actor.

If there is no mapped parent though, right now we just unset the implicit
grab actor. This leaves the grab in kind of a limbo state though: There is
no grab actor, but the grab continues to exist (because press_count > 0).
And when the grab later is ended on a button release, things blow up. To
fix it, let's avoid this limbo state and instead end the grab right away
when there is no mapped parent.