window: Allow placing new windows later
Window placement, even when "forced" in meta_window_force_placement() might
not happen in case the window is maximized, minimized, or fullscreen (see
place_window_if_needed()).

In that case we shouldn't set window->placed to TRUE, instead we should try
to place the window later when it's no longer maximized etc.

To allow for that, rework the placement code a bit so that window->placed
gets set from the actual placement functions. Because this changes
meta_window_force_placement() more into a meta_window_maybe_place() method
(it always kinda was, but now it's more obvious), rename the function to
that.