add a few assertions to catch a weird gesture crash
ClutterGesture crashed in handle_event() because for some reason a
TOUCHPAD_HOLD event made it through. This hold event had the core pointer as
its input device, and we already had a SeqData for a mouse point, so we
tried to use this seq data for the event and crashed in an assertion.

Backtrace:
    at ../clutter/clutter/clutter-main.c:597
    at ../clutter/clutter/clutter-stage-view.c:795