wayland/text-input: Replace focus for Aliendalvik with a proxy client
While there is wl_text_input v3 support in Aliendalvik, the implementation
is so broken it's impossible to use, even with workarounds in the
compositor.

Instead we don't use wl_text_input for Aliendalvik clients at all (we force
disable the protocol in libwayland) and trick Aliendalvik into using the
dbus based protocol that's used on Sailfish OS. We shim this dbus
based protocol in the alienkeyboard-maliit-shim. This shim also registers as
a wayland client for wl_text_input with us, and then acts as a drop-in for
wl_text_input of the Aliendalvik surface.

We therefore need a way to allow a separate Wayland client to handle text
input for a surface that belongs to a different client, clearly a risky
endeavour from a security perspective.

We implement it using a "magic string" kind of solution, where surrounding
text and commit string are abused for sending "protocol messages". This is
scary, because any app speaking wl_text_input could therefore become such a
proxy, so we take the following measures to somewhat lock this down:

- Only a single client is allowed to be the magic client at a time, any
subsequent tries to become the magic client trigger an error on the logs.
Ideally we should only allow setting the magic client once (possibly within
x seconds after the startup of the compositor). The shim on the other side
is obviously assumed to be started automatically here, probably this shim
should loudly complain via notification API in case it's unable to become
the magic client.

- Proxying of input is restricted to windows that have the "alien_" prefix
on their wm_class. Additionally those windows must not register a
wl_text_input client of themselves.