wayland/dma-buf: Use a guessed dmabuf stride for Aliendalvik
Aliendalvik will calculate the dmabuf stride once when creating the first
wayland surface, and then re-use that stride for every new buffer, even if
the buffer size changes. It does report a correct width/height for the new
buffer though, so we can guess the corrected stride and therefore fix the
failure to realize the buffer.