wayland: Spam keyboard enter events when Aliendalvik tries to minimize
Aliendalvik uses keyboard focus state for some kind of window activity
tracking, and that seems to be somewhat buggy: When switching window to
another Alien window, it detects the window as suspended and for some
reason sends input events to the Alien window that was focused previously.

When a window gets detected as suspended, Aliendalvik also tries to
minimize that window, so it seems like we can use that request and spam a
bunch of wl_keyboard::enter() events to the window to ensure the suspend
state gets tracked correctly again.