clutter/actor: Return empty box in get_transformed_extents() rather than NaN
When an actor doesn't have an allocation yet, its priv->allocation box is
set to CLUTTER_ACTOR_BOX_UNINITIALIZED, which means the origin and size of
the box are Infinity. When we input Infinity into our matrix calculations in
_clutter_actor_transform_and_project_box(), the results will be garbage NaN
values.

Rather than returning NaN values to the caller and having the caller deal
with that, lets return a 0-sized box in those cases. This still means the
caller has to deal with it and potentially think about the case, but at
least it won't have to do awkward checks for
clutter_actor_box_is_initialized() beforehand or even isNaN() checks
afterwards.