clutter: track implicit grab actors differently
I'm a bit scared there might be a fundamental problem with a behavior of
clutterGrabs that's incompatible with our tracking of implicit grab actors:

A clutterGrab can be on an actor that's completely outside the currently
picked event tree, and it'll still receive events. We kinda assume that the
implicit grab actor is always somewhere on that event tree though. This ends
up triggering an assertion in one super specific case: Using touch-hold to
move the cursor in the file rename popup in nautilus, and at the same time
triggering touch swipe-up to go into overview.