wayland-popup: Don't grab input from compositor for wayland popups
This is most likely not how grab=true in the popover protocol should
behavior, but it's quite annoying that you can't interact with the shell
during a wayland popup is visible. Apparently designers want this behavior
over the grabbing one anyway, but that message didn't seem to ever make it
through to gtk.