window: Avoid a division by 0 when new window has empty rect
We can crash here with a floating point exception right now. Reproducer is
to open the preferences window in cheese with mobile-shell in mobile (aka
maximized-windows) mode. Mobile shell moves the new window onto a new
workspace and then tries to activate that workspace, which in turn calls
meta_window_show() while the window still has a 0-sized rect.

Fix the crash by checking whether the area is 0 before we do our division.